Refactor party size to use long instead of int
Updated `JsonParty.Size` to `long[]?` to support larger values.
Replaced `CurrentSize` and `MaxSize` in `Party` with `PartySize`.
Introduced `PartySize` class for better encapsulation.
Updated `PartySizeProperties` to use `long` for size properties.

+++ NetCord/Gateway/JsonModels/JsonParty.cs
@@ -8,5 +8,5 @@ public class JsonParty
     public string? Id { get; set; }
 
     [JsonPropertyName("size")]
-    public int[]? Size { get; set; }
+    public long[]? Size { get; set; }
 }
